# public/openapi/v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Creditax.ai API",
    "version": "v2.1.0",
    "description": "API-first platform bridging tax compliance and creditworthiness for the Nigerian market. Naira-only pricing; every demo-build response is flagged `demo_seed: true`. Rate limiting: 100 requests/min per key (in-memory fixed window; a shared Redis store is the multi-instance upgrade). Re-verify tax figures against the current NTA/NTAA/FIRS schedule before production use.",
    "contact": { "name": "Creditax.ai Developer Portal", "url": "/developers" }
  },
  "servers": [{ "url": "https://api.creditax.ai" }],
  "security": [{ "apiKey": [] }],
...
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key issued via POST /api/v1/api-keys. Demo keys are deterministic and masked (`sk_live_****...****8f3a`)."
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" },
          "detail": { "type": "string" }
        }
      },
      "RateLimitExceeded": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string", "example": "rate_limit_exceeded" },
          "detail": { "type": "string" },
          "retry_after_sec": { "type": "integer" },
          "note": { "type": "string" },
          "demo_seed": { "type": "boolean" }
        },
        "description": "Returned with HTTP 429 when the 100 req/min window is exhausted; `Retry-After` header is also set."
      },
...
      "ChatRequest": {
        "type": "object",
        "required": ["question"],
        "properties": {
          "question": { "type": "string" },
          "locale": { "type": "string", "enum": ["en", "yo", "ha", "ig"], "default": "en" },
          "topK": { "type": "integer", "minimum": 1, "maximum": 8, "default": 5 },
          "conversationId": { "type": "string" },
          "userId": { "type": "string" },
          "tier": { "type": "string", "enum": ["free", "plus", "professional", "enterprise"] },
          "meteredAction": { "type": "string", "enum": ["chat", "calc", "upload", "report", "bvn"] },
          "docId": { "type": "string", "description": "Attached document pinned as context for this turn." },
          "loggedIn": { "type": "boolean" },
          "webhookUrl": { "type": "string", "format": "uri", "description": "Async delivery stub (always flagged delivered: false in Track A)." }
        }
      },
      "ChatResponse": {
        "type": "object",
        "description": "Grounded RAG answer with citations and confidence; multi-turn via conversationId.",
        "required": ["answer", "demo_seed"],
        "properties": {
          "answer": { "type": "string" },
          "citations": { "type": "array", "items": { "type": "object" } },
          "confidence": { "type": "number" },
          "conversationId": { "type": "string" },
          "webhook": { "type": "object" },
          "quota": { "type": "object", "description": "Meter result when the caller passed a user context; a 403 upgrade wall renders when a free-tier cap is hit." },
          "demo_seed": { "type": "boolean" }
        }
      },
      "CreditScoreResponse": {
        "type": "object",
        "description": "VantageScore-shaped snapshot; score = f(income, stability, savings rate, debt burden, tax compliance).",
        "required": ["score", "factors", "demo_seed"],
        "properties": {
          "score": { "type": "number" },
          "factors": { "type": "array", "items": { "type": "object" } },
          "simulator": { "type": "object" },
          "demo_seed": { "type": "boolean" }
        }
      },
      "ApiKeyRecord": {
        "type": "object",
        "required": ["id", "name", "key", "status", "demo_seed"],
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "key": { "type": "string", "description": "Masked: sk_live_****...****8f3a" },
          "environment": { "type": "string", "enum": ["live", "test"] },
          "scopes": { "type": "array", "items": { "type": "string" } },
          "status": { "type": "string", "enum": ["active", "revoked"] },
          "usage": { "type": "number", "description": "Live counter, bumped on each admitted request by the rate limiter." },
          "limit": { "type": "number" },
          "last_used": { "type": "string" },
          "demo_seed": { "type": "boolean" }
        }
      },
      "WebhookEndpoint": {
        "type": "object",
        "required": ["id", "url", "events", "demo_seed"],
        "properties": {
          "id": { "type": "string" },
          "url": { "type": "string", "format": "uri" },
          "events": { "type": "array", "items": { "type": "string" } },
          "active": { "type": "boolean" },
          "secret": { "type": "string" },
          "demo_seed": { "type": "boolean" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/credit/score": {
      "get": {
        "tags": ["Credit"],
        "operationId": "getCreditScore",
        "summary": "VantageScore-shaped snapshot (filing streak beats bank balance)",
        "responses": {
          "200": { "description": "Score + factor breakdown.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreditScoreResponse" } } } },
          "429": { "description": "Rate limit window exhausted.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RateLimitExceeded" } } } }
        }
      }
    },
    "/api/v1/chat": {
      "post": {
        "tags": ["Chat (RAG)"],
        "operationId": "ragChat",
        "summary": "Grounded Nigerian tax-law answer with citations and confidence",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ChatRequest" } } } },
        "responses": {
          "200": { "description": "Answer + citations + confidence + quota meter result.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ChatResponse" } } } },
          "400": { "description": "Missing question / invalid JSON.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "403": { "description": "Quota cap hit — upgrade wall payload on free-tier exhaustion." }
        }
      }
    },
    "/api/v1/api-keys": {
      "get": {
        "tags": ["Developer"],
        "operationId": "listApiKeys",
        "summary": "List API keys (demo-seeded on first access)",
        "responses": { "200": { "description": "Key records.", "content": { "application/json": { "schema": { "type": "object", "properties": { "keys": { "type": "array", "items": { "$ref": "#/components/schemas/ApiKeyRecord" } }, "demo_seed": { "type": "boolean" } } } } } } }
      },
      "post": {
        "tags": ["Developer"],
        "operationId": "createApiKey",
        "summary": "Create a key",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["name"], "properties": { "name": { "type": "string" }, "environment": { "type": "string", "enum": ["live", "test"] }, "scopes": { "type": "array", "items": { "type": "string" } }, "limit": { "type": "number" } } } } } },
        "responses": { "200": { "description": "Created record.", "content": { "application/json": { "schema": { "type": "object", "properties": { "key": { "$ref": "#/components/schemas/ApiKeyRecord" }, "demo_seed": { "type": "boolean" } } } } } }, "400": { "description": "Missing name.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } } }
      }
    },
...
    "/api/v1/webhooks": {
      "get": {
        "tags": ["Developer"],
        "operationId": "listWebhooks",
        "summary": "List webhook endpoints",
        "responses": { "200": { "description": "Endpoint registry.", "content": { "application/json": { "schema": { "type": "object", "properties": { "endpoints": { "type": "array", "items": { "$ref": "#/components/schemas/WebhookEndpoint" } }, "demo_seed": { "type": "boolean" } } } } } } }
      },
      "post": {
        "tags": ["Developer"],
        "operationId": "createWebhook",
        "summary": "Register a webhook endpoint",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["url"], "properties": { "url": { "type": "string", "format": "uri" }, "userId": { "type": "string" }, "events": { "type": "array", "items": { "type": "string" } } } } } } },
        "responses": { "200": { "description": "Registered endpoint.", "content": { "application/json": { "schema": { "type": "object", "properties": { "endpoint": { "$ref": "#/components/schemas/WebhookEndpoint" } } } } } } }
      }
    },
...
    "/api/v1/documents": {
      "get": {
        "tags": ["Platform"],
        "operationId": "listDocuments",
        "summary": "List a user's processed documents (lazily seeded for demo users)",
        "parameters": [{ "name": "userId", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Document records." } }
      }
    },
...
  }
}